+++ src/main/java/com/birthday/aisen/mapper/UserMapper.java
@@ -5,6 +5,6 @@ import com.birthday.aisen.entity.User;
 import java.util.List;
 
 public interface UserMapper {
-    User getUserById(int uid);
+    User getUserById(long uid);
     List<User> getUsersByName(String username);
 }
