+++ src/main/java/com/birthday/aisen/service/IUserService.java
@@ -5,7 +5,7 @@ import com.birthday.aisen.dto.UserDTO;
 import java.util.List;
 
 public interface IUserService {
-    public UserDTO getUserById(int uid);
+    public UserDTO getUserById(long uid);
 
     public List<UserDTO> getUsersByName(String username);
 }
